rename some packages and update imports
