+++ tests/test_interpreter.py
@@ -1,6 +1,6 @@
 import unittest
 from src.interpreter import Interpreter
-from src.parser_expression import Binary, Literal, Unary
+from src.ast_node_expression import Binary, Literal, Unary
 from src.tokens import TokenOsu, TokenType
